requirements.txt
numpy
tensorflow
ipython
django
Pillow